Massive refactor

+++ platformio_nuli_avionics_flight_software/src/drivers/arduino/ArduinoVoltageSensor.cpp
@@ -1,7 +1,7 @@
 #include "ArduinoVoltageSensor.h"
 #include "Arduino.h"
 
-ArduinoVoltageSensor::ArduinoVoltageSensor(uint8_t pin, double scaleFactor) : VoltageSensor(scaleFactor), m_pin(pin) {}
+ArduinoVoltageSensor::ArduinoVoltageSensor(uint8_t pin, float scaleFactor) : VoltageSensor(scaleFactor), m_pin(pin) {}
 
 
 void ArduinoVoltageSensor::read() {
